Command line option for icon theme
Without adding a config file, theme selection is currently done via a
command line switch. The `--icon-theme` parameter allows selecting the
icon theme by specifying a directory name.

The theme name is sanitized to prevent any path traversal attacks. Only
the characters a to z are allowed. This can be extened if needed.

The specified theme is tried first. If it is not found, the theme named
`default` will be selected as a fallback.